added user and category update Not finished

+++ src/app/admin/users/users.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import Swal from 'sweetalert2';
 import { AdminService, USER } from '../admin.service';
 
 @Component({
@@ -16,10 +17,24 @@ export class UsersComponent implements OnInit {
   }
 
   onDelete(user: USER) {
-    this.adminservice
-      .deleteUser(user)
-      .subscribe(
-        () => (this.users = this.users.filter((t) => t.id !== user.id))
-      );
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.adminservice
+          .deleteUser(user)
+          .subscribe(
+            () => (this.users = this.users.filter((t) => t.id !== user.id))
+          );
+        Swal.fire('Deleted!', 'Your file has been deleted.', 'success');
+        this.users = this.users.filter((t) => t.id !== user.id);
+      }
+    });
   }
 }
